pyproject.toml
[tool.poetry]
name = "bigdataproject"
version = "0.1.0"
description = ""
authors = ["Sumeet_Sule <sule.sumeet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
pandas = "^2.2.3"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
